I fucking HATE CORS, TWO HOURS ON THIS PIECE OF SHIT
